bot/pyproject.toml
[project]
name = "yandex-music-downloader-bot"
version = "0.1.0"
description = "Integration yandex music with telegram!"
authors = [
    {name = "sht0rmx", email = "mail@sht0rmx.ru"},
]
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "aiogram>=3.0",
    "sqlmodel<1.0.0,>=0.0.21",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19",
    "python-dotenv",
    "python-decouple",
    "watchdog>=6.0.0",
    "yandex-music>=2.2.0",
    "requests>=2.32.5",
    "odesli>=0.0.2",
    "mutagen>=1.47.0",
]

[project.urls]
"Homepage" = "https://github.com/sht0rmx/yandmdbot"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project.optional-dependencies]
dev = [
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
]

[tool.mypy]
strict = true
python_version = "3.12"
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py312"
exclude = ["venv", ".venv"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused args
    "T201",   # print statements
]
ignore = [
    "E501",  # handled by black
    "B008",  # function calls in arg defaults
    "W191",  # tabs
    "B904",  # raise Exception without from
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
branch = true
dynamic_context = "test_function"
